do not flush when not DEBUG

+++ flush.py
@@ -1,8 +1,11 @@
 from basic.database_wrapper_redis import DatabaseWrapperRedis
 import bot_config
 
-DATABASE = DatabaseWrapperRedis(
-    host=bot_config.DB_HOST, port=bot_config.DB_PORT,
-    db=bot_config.DB_NUM)
+if bot_config.DEBUG:
+    DATABASE = DatabaseWrapperRedis(
+        host=bot_config.DB_HOST, port=bot_config.DB_PORT,
+        db=bot_config.DB_NUM)
 
-DATABASE.flushdb()
+    DATABASE.flushdb()
+else:
+    print('It is in production mode. Did you really mean to flush it?')
